# docs/openapi/c2c-market-api-v1.yaml
openapi: 3.1.0
info:
  title: C2CMarket API
  version: 0.1.0
  summary: Backend contract foundation and official price lead slice.
  description: |
    This contract covers the first executable backend slice. C2CMarket is a price intelligence
    and community matching platform. It does not process payments, escrow funds, store third-party
    credentials, store plaintext native passwords, auto-deliver keys/accounts, or proxy upstream API
    traffic. Native C2CMarket account passwords are stored only as salted hashes.
servers:
  - url: http://localhost:8080
security:
  - sessionCookie: []
paths:
...
  /api/v1/me/api-purchase-intents/{id}/orders:
    post:
      operationId: createMyAPIOrder
      tags: [APIOrders]
      security:
        - sessionCookie: []
          csrfToken: []
      summary: Create an API service order from an existing purchase intent.
      description: Creates one explicit API order for the purchase intent. The request only selects a payment method; amount, currency, service title, counterparties, payment window, and private payment instructions are frozen server-side. The platform does not process payment, escrow funds, proxy API traffic, store credentials, or automatically deliver API keys. A purchase intent can create at most one API order; duplicate or concurrent order creation returns `409 API_PURCHASE_INTENT_HAS_ORDER`.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPIOrderRequest"
      responses:
        "201":
          description: API order created. Payment instructions are not included in this response.
          headers:
            Cache-Control:
              schema:
                type: string
              description: Always `private, no-store`.
            Location:
              schema:
                type: string
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIOrder"
        "409":
          description: Conflict. Duplicate order creation for the same purchase intent returns Problem `code` `API_PURCHASE_INTENT_HAS_ORDER`.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          $ref: "#/components/responses/Problem"
  /api/v1/me/api-orders:
    get:
      operationId: listMyAPIOrders
      tags: [APIOrders]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Current user's API orders as buyer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIOrderList"
  /api/v1/me/api-orders/{id}:
    get:
      operationId: getMyAPIOrder
      tags: [APIOrders]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Buyer API order detail. Sensitive order details must be served with Cache-Control private, no-store.
          headers:
            Cache-Control:
              schema:
                type: string
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIOrder"
        "404":
          $ref: "#/components/responses/Problem"
...
  /api/v1/owner/api-orders/{id}:
    get:
      operationId: getOwnerAPIOrder
      tags: [OwnerAPIOrders]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Owner API order detail. Sensitive order details must be served with Cache-Control private, no-store.
          headers:
            Cache-Control:
              schema:
                type: string
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIOrder"
        "404":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: c2c_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Defaults to 20 and must be between 1 and 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque pagination cursor returned as nextCursor. Current implementation encodes an offset cursor and clients must not inspect it.
      schema:
        type: string
...
  responses:
    Problem:
      description: Problem Details error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
  schemas:
    DecimalString:
      type: string
      pattern: '^[0-9]+(\.[0-9]{1,8})?$'
      maxLength: 32
...
    CreateAPIOrderRequest:
      type: object
      additionalProperties: false
      required: [paymentMethod]
      properties:
        paymentMethod:
          type: string
          enum: [wechat, alipay, usdt]
          description: Selected payment method. The server freezes all pricing, party, window, and instruction snapshots.
...
    APIOrder:
      type: object
      additionalProperties: false
      description: API service order. The order stores fulfillment status and off-platform action summaries only; C2CMarket does not process payments, provide escrow, proxy API calls, store credentials, or automatically deliver API keys.
      required:
        - id
        - apiPurchaseIntentId
        - apiServiceId
        - status
        - disputeStatus
        - serviceTitleSnapshot
        - serviceVersionSnapshot
        - billingModeSnapshot
        - amount
        - currency
        - selectedPaymentMethod
        - paymentWindowMinutesSnapshot
        - paymentExpiresAt
        - version
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        apiPurchaseIntentId:
          type: string
          format: uuid
        apiServiceId:
          type: string
          format: uuid
        buyerUserId:
          type: string
          format: uuid
          description: Present on owner-side views.
        sellerUserId:
          type: string
          format: uuid
          description: Present on buyer-side views.
        status:
          type: string
          enum: [pending_payment, payment_submitted, paid_confirmed, delivery_submitted, completed, cancelled]
        disputeStatus:
          type: string
          enum: [none, open, closed]
        disputeCaseId:
          type: string
          format: uuid
        serviceTitleSnapshot:
          type: string
        serviceVersionSnapshot:
          type: integer
        billingModeSnapshot:
          type: string
        selectedPackageId:
          type: string
          format: uuid
        selectedPackageSnapshot:
          type: string
        quoteVersionSnapshot:
          type: integer
        amount:
          $ref: "#/components/schemas/DecimalString"
        currency:
          type: string
          enum: [CNY, USDT]
        selectedPaymentMethod:
          type: string
          enum: [wechat, alipay, usdt]
        paymentWindowMinutesSnapshot:
          type: integer
          minimum: 3
          maximum: 15
        paymentExpiresAt:
          type: string
          format: date-time
        paymentSummary:
          type: string
        paymentSubmittedAt:
          type: [string, "null"]
          format: date-time
        paidConfirmedAt:
          type: [string, "null"]
          format: date-time
        deliveryNote:
          type: string
          description: Pure text off-platform delivery prompt only; Authorization headers, API keys, tokens, passwords, cookies, JWTs, subscription links, proxy node links, credentials, and attachments are rejected.
        deliverySubmittedAt:
          type: [string, "null"]
          format: date-time
        completedAt:
          type: [string, "null"]
          format: date-time
        cancelledAt:
          type: [string, "null"]
          format: date-time
        cancelReason:
          type: string
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    APIOrderList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/APIOrder"
        nextCursor:
          type: [string, "null"]
...
    ProblemDetails:
      type: object
      required: [type, title, status, code, detail, instance, requestId]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string